Fix error in 'used' quota when creating share
The 'Create Share' button calls tenant_quota_usages() to find out if a
user has sufficient quota, disabling the button if there is insufficient
quota. However, the call tenant_quota_usages() is not indempotent,
resulting in usages showing the wrong value.

This is because usages.tally() increments 'used' by value passed to it,
instead of setting it to the value. To work around this, we set 'used'
to zero before calling tally() so that this function is idempotent.

Change-Id: I3193adc95cebb29f13afa6ebd848985495d60d83